Reverse synstack before setting commentstring
I had a problem that the commentstring does not get set correctly in the
following case in a JSX file (cursor noted by |):

```jsx
const MyComponent = () => {
    return {
        <button
            onClick={() => {
                |console.log('hello')
                 ^ Cursor is here
            }}
        >
            Hello!
        </button>
    }
}
```

The commentstring at that location should be `// %s` but it's not really
possible to match the syntax group correctly. Here's how the synstack
looks like at that location:

['jsFuncBlock', 'jsParen', 'jsxRegion', 'jsxTag',
'jsxEscapeJsAttributes', 'jsFuncBlock', 'jsGlobalObjects']

When looping through this list from left to right, the `jsxRegion`
syntax group gets matched first and the commentstring will be
`{/*%s*/}`. If the synstack is reversed, however, the first match will
be `jsFuncBlock` and the commentstring will be set to `// %s` as
expected.