Dynamically set comments setting
The problem is that, in Vue files, the `comments` setting is set to
support HTML comments. This means that if you have enabled `r` in
`formatoptions` and you press <CR> while in a comment in the JS section,
then the comment character won't get auto-inserted.

Since the `commentstring` variable is already being set dynamically with
the same logic, then it makes sense to support the `comments` option
here as well.