Dynamically set comments setting
I had a problem where Vim doesn't automatically insert the comment
character to the start of the line when I press <CR> in a comment in Vue
files even if I have the `r` option in my `formatoptions`.

The problem turned out to be because Vue files have the `comments`
setting set to `s:<!--,m:    ,e:-->` which is valid for HTML, but not
for the sections that include other languages. The `comments` setting
should be modified automatically in the exact same way that
`commentstring` is currently. So, I thought that it might make sense to
add this functionality to this plugin.